1.新增用户角色权限控制， 新增访问用户权限 2.删除子任务描述 3.管理员用户新增用户功能 4.确保手工终止任务不会因为处理器down机而无法终止

+++ xxl-job-admin/src/main/java/com/xxl/job/admin/controller/JobInfoController.java
@@ -1,23 +1,29 @@
 package com.xxl.job.admin.controller;
 
-import com.xxl.job.admin.core.enums.ExecutorFailStrategyEnum;
-import com.xxl.job.admin.core.model.XxlJobGroup;
-import com.xxl.job.admin.core.model.XxlJobInfo;
-import com.xxl.job.admin.core.route.ExecutorRouteStrategyEnum;
-import com.xxl.job.admin.dao.XxlJobGroupDao;
-import com.xxl.job.admin.service.XxlJobService;
-import com.xxl.job.core.biz.model.ReturnT;
-import com.xxl.job.core.enums.ExecutorBlockStrategyEnum;
-import com.xxl.job.core.glue.GlueTypeEnum;
+import java.util.List;
+import java.util.Map;
+
+import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import javax.annotation.Resource;
-import java.util.List;
-import java.util.Map;
+import com.xxl.job.admin.controller.interceptor.PermissionInterceptor;
+import com.xxl.job.admin.core.enums.ExecutorFailStrategyEnum;
+import com.xxl.job.admin.core.model.XxlJobGroup;
+import com.xxl.job.admin.core.model.XxlJobInfo;
+import com.xxl.job.admin.core.route.ExecutorRouteStrategyEnum;
+import com.xxl.job.admin.core.util.CookieUtil;
+import com.xxl.job.admin.dao.XxlJobGroupDao;
+import com.xxl.job.admin.service.UserRoleService;
+import com.xxl.job.admin.service.XxlJobService;
+import com.xxl.job.core.biz.model.ReturnT;
+import com.xxl.job.core.enums.ExecutorBlockStrategyEnum;
+import com.xxl.job.core.glue.GlueTypeEnum;
 
 /**
  * index controller
@@ -27,13 +33,21 @@ import java.util.Map;
 @RequestMapping("/jobinfo")
 public class JobInfoController {
 
+
 	@Resource
 	private XxlJobGroupDao xxlJobGroupDao;
 	@Resource
 	private XxlJobService xxlJobService;
+
+	@Resource
+	private UserRoleService userRoleService;
+	
 	
 	@RequestMapping
-	public String index(Model model, @RequestParam(required = false, defaultValue = "-1") int jobGroup) {
+	public String index(HttpServletRequest request, Model model, @RequestParam(required = false, defaultValue = "-1") int jobGroup) {
+		String userName = CookieUtil.getValue(request, PermissionInterceptor.LOGIN_USER_NAME);
+		model.addAttribute("userName", userName);
+		boolean isEditable = isEditable(request);
 
 		// 枚举-字典
 		model.addAttribute("ExecutorRouteStrategyEnum", ExecutorRouteStrategyEnum.values());	// 路由策略-列表
@@ -45,10 +59,17 @@ public class JobInfoController {
 		List<XxlJobGroup> jobGroupList =  xxlJobGroupDao.findAll();
 		model.addAttribute("JobGroupList", jobGroupList);
 		model.addAttribute("jobGroup", jobGroup);
+		
+		model.addAttribute("editable", isEditable);
 
 		return "jobinfo/jobinfo.index";
 	}
 	
+	private boolean isEditable(HttpServletRequest request) {
+		String userName = CookieUtil.getValue(request, PermissionInterceptor.LOGIN_USER_NAME);
+		return userRoleService.isUserAsAdmin(userName);
+	}
+
 	@RequestMapping("/pageList")
 	@ResponseBody
 	public Map<String, Object> pageList(@RequestParam(required = false, defaultValue = "0") int start,  
@@ -60,38 +81,62 @@ public class JobInfoController {
 	
 	@RequestMapping("/add")
 	@ResponseBody
-	public ReturnT<String> add(XxlJobInfo jobInfo) {
-		return xxlJobService.add(jobInfo);
+	public ReturnT<String> add(HttpServletRequest request, XxlJobInfo jobInfo) {
+		if(isEditable(request)) {
+			return xxlJobService.add(jobInfo);
+		}else {
+			return new ReturnT<String>(ReturnT.FAIL_CODE, "没有操作权限");
+		}
 	}
 	
 	@RequestMapping("/reschedule")
 	@ResponseBody
-	public ReturnT<String> reschedule(XxlJobInfo jobInfo) {
-		return xxlJobService.reschedule(jobInfo);
+	public ReturnT<String> reschedule(HttpServletRequest request, XxlJobInfo jobInfo) {
+		if(isEditable(request)) {
+			return xxlJobService.reschedule(jobInfo);
+		}else {
+			return new ReturnT<String>(ReturnT.FAIL_CODE, "没有操作权限");
+		}
 	}
 	
 	@RequestMapping("/remove")
 	@ResponseBody
-	public ReturnT<String> remove(int id) {
-		return xxlJobService.remove(id);
+	public ReturnT<String> remove(HttpServletRequest request, int id) {
+		if(isEditable(request)) {
+			return xxlJobService.remove(id);
+		}else {
+			return new ReturnT<String>(ReturnT.FAIL_CODE, "没有操作权限");
+		}
 	}
 	
 	@RequestMapping("/pause")
 	@ResponseBody
-	public ReturnT<String> pause(int id) {
-		return xxlJobService.pause(id);
+	public ReturnT<String> pause(HttpServletRequest request, int id) {
+		if(isEditable(request)) {
+			return xxlJobService.pause(id);
+		}else {
+			return new ReturnT<String>(ReturnT.FAIL_CODE, "没有操作权限");
+		}
 	}
 	
 	@RequestMapping("/resume")
 	@ResponseBody
-	public ReturnT<String> resume(int id) {
-		return xxlJobService.resume(id);
+	public ReturnT<String> resume(HttpServletRequest request, int id) {
+		if(isEditable(request)) {	
+			return xxlJobService.resume(id);
+		}else {
+			return new ReturnT<String>(ReturnT.FAIL_CODE, "没有操作权限");
+		}
 	}
 	
 	@RequestMapping("/trigger")
 	@ResponseBody
-	public ReturnT<String> triggerJob(int id) {
-		return xxlJobService.triggerJob(id);
+	public ReturnT<String> triggerJob(HttpServletRequest request, int id) {
+		if(isEditable(request)) {
+			return xxlJobService.triggerJob(id);
+		}else {
+			return new ReturnT<String>(ReturnT.FAIL_CODE, "没有操作权限");
+		}
 	}
 	
 }
